\name{O.mykiss}

\alias{O.mykiss}

\docType{data}

\title{Test data from a 21 day fish test}

\description{
  Test data from a 21 day fish test following the guidelines OECD GL204, 
  using the test organism \emph{Oncorhynchus mykiss}.
}

\usage{data(O.mykiss)}

\format{
  A data frame with 70 observations on the following 2 variables.
  \describe{
    \item{\code{conc}}{a numeric vector of concentrations (mg/l)}
    \item{\code{weight}}{a numeric vector of wet weights (g)}
  }
}

\details{
  Weights are measured after 28 days.
}

\source{
  Organisation for Economic Co-operation and Development (OECD) (2006)
  \emph{CURRENT APPROACHES IN THE STATISTICAL ANALYSIS OF ECOTOXICITY DATA: A GUIDANCE TO APPLICATION - ANNEXES},
  Paris (p. 65).
}

%\references{}

\examples{

O.mykiss

}
\keyword{datasets}
