\name{RGRcurve}

\alias{RGRcurve}

\docType{data}

\title{Relative growth rates}

\description{
  Relative growth rates
}

\usage{data(RGRcurve)}

\format{
  A data frame with 41 observations on the following 2 variables.
  \describe{
    \item{\code{Day}}{a numeric vector denoting the number of days elapsed}
    \item{\code{RGR}}{a numeric vector of relative growth rates}
  }
}

\details{
  Data stem from a large experiment on the effect of herbicides on aquatic plants.
}

\source{
  Cedergreen, N. and Andersen, L. and Olesen, C. F. and Spliid, H. H. and Streibig, J. C. (2005)
  Does the effect of herbicide pulse exposure on aquatic plants depend on K-ow or mode of action?,
  \emph{Aquatic Toxicology}, \bold{72}, 261--271.
}

%\references{}

\examples{

RGRcurve

}
\keyword{datasets}
