\name{sockeye}

\alias{sockeye}

\docType{data}

\title{Data on stock and recruitment of sockeye salmon in Skeena River}

\description{
  Spawning stock and resulting recruitment of sockeye salmon were recorded over the
  years 1940-1967 in Skeena River.
}

\usage{data(sockeye)}

\format{
  A data frame with 28 observations on the following 3 variables.
  \describe{
    \item{\code{year}}{a numeric vector of years}
    \item{\code{spawners}}{a numeric vector giving number of spawning fish (in thousands)}
    \item{\code{recruits}}{a numeric vector containing the recruitment (thousands)}
  }
}

\details{
  The dataset is usually analysed without case 12 because a rockslide occurred that year.
}

\source{
  Carroll, R. J. and Ruppert, D. (1988) \emph{Transformation and Weighting in Regression}, 
  New York: Chapman and Hall (p. 140).
}

%\references{}

\examples{
}
\keyword{datasets}
